Carregamento de imagem para o feed

+++ Zup/app/src/main/java/school/sptech/zup/model/EventoFeed.kt
@@ -1,3 +1,5 @@
 package school.sptech.zup.model
 
-data class EventoFeed(val Titulo: String, val Image: Byte, val Descricao: String)
+import java.net.URL
+
+data class EventoFeed(val Titulo: String, val Image: Int, val Descricao: String)
